[SYSTEMDS-3404] Synchronous w/ Backup Worker for ParamServer
This adds a new update type to the parameter server implementation. This update type is similar to the synchronous mode, but ignores some stragglers results, with the benefit of starting the next iteration early for the other workers. It is called SBP, referring to "Synchronous with Backup-workers Parallel".

There are currently still a few limitations:
1. Stragglers are not stopped early, their result is thrown away when they return.
2. There is no balancing mechanism in place for unbalanced federated workloads. The largest federated partition will probably be thrown away every time.
We can think about fixing some of those limitations or leave it up to the end user.

Closes #1653