[SYSTEMDS-3453] CLA Offsets append
This commit adds support for offset lists to append in all cases.
Also added is parallelization of appendN for all column groups in
a read in compressed matrix.
Further specialized formats for appendN is added for OffsetByte,
that is the offset type most in need of specialization, since if
there is a high number of offsets, the encoding ends up using offset
byte. It is also beneficial since this encoding benefit from java
vectorization when copying byte arrays using system calls.