[SYSTEMDS-3413] Row/Col aggregation for countDistinct
This patch converts countDistinct() from  a non-parameterized builtin to
a parameterized builtin function to allow for 1 new parameter: dir for
direction. The value of dir can be r and c, denoting row-wise and
column-wise aggregation respectively. This patch only implements CP and
the SP case will throw a NotImplementedException()- the latter case will
be addressed in a subsequent patch.

Closes #1677