[SYSTEMDS-3461] FrameBlocks Arrays Separation
This commit removes the tight binding of the underlying
arrays in frame block to its arrays for columns value types.
There is one binding missing that is to pipe frames to python.
This will be refactored in a following commit.

This further fixes the indentation of one of the frame test files
(tabs vs spaces) and moves the functional frame tests that was
placed in component tests to frames.