[SYSTEMDS-3473] Push down rmvars for asynchronous instructions
This patch repositions the rmvar instructions for the inputs to an
asynchronous instruction after the consumers of the asynchronous
instruction. This change allows keeping the inputs to an asynchronous
operator alive until get() is called for the future object.
Moreover, this patch adds more tests for the new operator ordering
and fixes minor bugs.

Closes #1745